requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0 